requirements.txt
asgiref==3.5.1
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
dj-database-url==1.2.0
Django==4.0.4
django-crispy-forms==2.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
gunicorn==20.1.0
psycopg2==2.9.5
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.4.0
XlsxWriter==3.0.8
xlwt==1.3.0
